refactor: Move MAX_WAIT_TIME_NOT_SET
The MAX_WAIT_TIME_NOT_SET belong to the TransferConstraint class, not
the ConstrainedTransfer class.
