+++ src/test/java/org/opentripplanner/model/transfer/TransferConstraintTest.java
@@ -3,7 +3,7 @@ package org.opentripplanner.model.transfer;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
-import static org.opentripplanner.model.transfer.ConstrainedTransfer.MAX_WAIT_TIME_NOT_SET;
+import static org.opentripplanner.model.transfer.TransferConstraint.MAX_WAIT_TIME_NOT_SET;
 import static org.opentripplanner.model.transfer.TransferPriority.ALLOWED;
 import static org.opentripplanner.model.transfer.TransferPriority.PREFERRED;
 
@@ -14,10 +14,18 @@ public class TransferConstraintTest implements TransferTestData {
 
   public static final int MAX_WAIT_TIME_ONE_HOUR = DurationUtils.duration("1h");
 
-  private final TransferConstraint NO_CONSTRAINS = new TransferConstraint(ALLOWED, false, false, MAX_WAIT_TIME_NOT_SET);
-  private final TransferConstraint RECOMMENDED = new TransferConstraint(TransferPriority.RECOMMENDED, false, false, MAX_WAIT_TIME_NOT_SET);
-  private final TransferConstraint STAY_SEATED = new TransferConstraint(ALLOWED, true, false, MAX_WAIT_TIME_NOT_SET);
-  private final TransferConstraint GUARANTIED = new TransferConstraint(ALLOWED, false, true, MAX_WAIT_TIME_NOT_SET);
+  private final TransferConstraint NO_CONSTRAINS = new TransferConstraint(
+          ALLOWED, false, false, MAX_WAIT_TIME_NOT_SET
+  );
+  private final TransferConstraint RECOMMENDED = new TransferConstraint(
+          TransferPriority.RECOMMENDED, false, false, MAX_WAIT_TIME_NOT_SET
+  );
+  private final TransferConstraint STAY_SEATED = new TransferConstraint(
+          ALLOWED, true, false, MAX_WAIT_TIME_NOT_SET
+  );
+  private final TransferConstraint GUARANTIED = new TransferConstraint(
+          ALLOWED, false, true, MAX_WAIT_TIME_NOT_SET
+  );
   private final TransferConstraint MAX_WAIT_TIME = new TransferConstraint(ALLOWED, false, false, MAX_WAIT_TIME_ONE_HOUR);
   private final TransferConstraint EVERYTHING = new TransferConstraint(PREFERRED, true, true, MAX_WAIT_TIME_ONE_HOUR);
