+++ src/test/java/org/opentripplanner/model/transfer/TransferServiceTest.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.model.transfer;
 
 import static org.junit.Assert.assertEquals;
-import static org.opentripplanner.model.transfer.ConstrainedTransfer.MAX_WAIT_TIME_NOT_SET;
+import static org.opentripplanner.model.transfer.TransferConstraint.MAX_WAIT_TIME_NOT_SET;
 import static org.opentripplanner.model.transfer.TransferPriority.ALLOWED;
 
 import java.util.List;
@@ -55,7 +55,9 @@ public class TransferServiceTest implements TransferTestData {
 
 
     ConstrainedTransfer transfer(TransferPoint from, TransferPoint to) {
-        var c = new TransferConstraint(ALLOWED, false, false, MAX_WAIT_TIME_NOT_SET);
+        var c = new TransferConstraint(
+                ALLOWED, false, false, MAX_WAIT_TIME_NOT_SET
+        );
         return new ConstrainedTransfer(from, to, c);
     }
 }
